requirements.txt
notion-client==2.2.1
requests>=2.31.0
beautifulsoup4>=4.12.2
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
selenium
python-dotenv